🔒 Warden: Fix unsound unsafe env var mutation in tests
Consolidate the test utility `EnvGuard` into a single module `test_utils::env_guard`
so that tests across the `autumn` crate share the exact same `ENV_LOCK` Mutex.
Previously, each test module defined its own `EnvGuard` and lock, which led to
data races and undefined behavior when running tests concurrently because
`std::env::set_var` is inherently unsafe and was being modified simultaneously.
